add workspace arg to cmd list
